+++ api/Controllers/PermissionController.cs
@@ -1,3 +1,4 @@
+using api.Authorization;
 using api.Interfaces;
 using api.Dtos.Permission;
 using api.Models;
@@ -12,24 +13,38 @@ namespace api.Controllers
     public class PermissionController : ControllerBase
     {
         private readonly IPermissionRepository _permissionRepository;
+        private readonly IPermissionGuard _guard;
 
-        public PermissionController(IPermissionRepository permissionRepository)
+        public PermissionController(IPermissionRepository permissionRepository, IPermissionGuard guard)
         {
             _permissionRepository = permissionRepository;
+            _guard = guard;
         }
 
         // GET: api/permissions
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<PermissionDto>>> GetPermissions()
+        public async Task<ActionResult<IEnumerable<PermissionDto>>> GetPermissions([FromQuery] Guid callerId)
         {
+            var authorization = await _guard.RequireAsync(callerId, Permissions.BackOfficeRead);
+            if (!authorization.Ok)
+            {
+                return authorization.ToActionResult();
+            }
+
             var permissions = await _permissionRepository.GetAllPermissionsAsync();
             return Ok(permissions.Select(p => p.ToPermissionDto()));
         }
 
         // GET: api/permissions/{id}
         [HttpGet("{id}")]
-        public async Task<ActionResult<PermissionDto>> GetPermission([FromRoute] Guid id)
+        public async Task<ActionResult<PermissionDto>> GetPermission([FromRoute] Guid id, [FromQuery] Guid callerId)
         {
+            var authorization = await _guard.RequireAsync(callerId, Permissions.BackOfficeRead);
+            if (!authorization.Ok)
+            {
+                return authorization.ToActionResult();
+            }
+
             var permission = await _permissionRepository.GetPermissionByIdAsync(id);
             if (permission == null)
             {
@@ -40,8 +55,14 @@ namespace api.Controllers
 
         // POST: api/permissions
         [HttpPost]
-        public async Task<ActionResult<PermissionDto>> CreatePermission([FromBody] CreatePermissionRequestDto permissionDto)
+        public async Task<ActionResult<PermissionDto>> CreatePermission([FromQuery] Guid callerId, [FromBody] CreatePermissionRequestDto permissionDto)
         {
+            var authorization = await _guard.RequireAsync(callerId, Permissions.UserManagement);
+            if (!authorization.Ok)
+            {
+                return authorization.ToActionResult();
+            }
+
             var permission = new Permission
             {
                 Code = permissionDto.Code,
@@ -54,8 +75,14 @@ namespace api.Controllers
 
         // PUT: api/permissions/{id}
         [HttpPut("{id}")]
-        public async Task<ActionResult<PermissionDto>> UpdatePermission([FromRoute] Guid id, [FromBody] UpdatePermissionRequestDto permissionDto)
+        public async Task<ActionResult<PermissionDto>> UpdatePermission([FromRoute] Guid id, [FromQuery] Guid callerId, [FromBody] UpdatePermissionRequestDto permissionDto)
         {
+            var authorization = await _guard.RequireAsync(callerId, Permissions.UserManagement);
+            if (!authorization.Ok)
+            {
+                return authorization.ToActionResult();
+            }
+
             var permission = new Permission
             {
                 Code = permissionDto.Code,
@@ -73,8 +100,14 @@ namespace api.Controllers
 
         // DELETE: api/permissions/{id}
         [HttpDelete("{id}")]
-        public async Task<ActionResult> DeletePermission([FromRoute] Guid id)
+        public async Task<ActionResult> DeletePermission([FromRoute] Guid id, [FromQuery] Guid callerId)
         {
+            var authorization = await _guard.RequireAsync(callerId, Permissions.UserManagement);
+            if (!authorization.Ok)
+            {
+                return authorization.ToActionResult();
+            }
+
             var deleted = await _permissionRepository.DeletePermissionAsync(id);
             if (!deleted)
             {
